Update to SilverStripe 3.3, fix a deprecation warning.

+++ framework/forms/gridfield/GridFieldExportButton.php
@@ -24,7 +24,7 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 	 * @var boolean
 	 */
 	protected $csvHasHeader = true;
-	
+
 	/**
 	 * Fragment to write the button to
 	 */
@@ -44,10 +44,10 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 	 */
 	public function getHTMLFragments($gridField) {
 		$button = new GridField_FormAction(
-			$gridField, 
-			'export', 
+			$gridField,
+			'export',
 			_t('TableListField.CSVEXPORT', 'Export to CSV'),
-			'export', 
+			'export',
 			null
 		);
 		$button->setAttribute('data-icon', 'download-csv');
@@ -103,8 +103,6 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 			? $this->exportColumns
 			: singleton($gridField->getModelClass())->summaryFields();
 		$fileData = '';
-		$columnData = array();
-		$fieldItems = new ArrayList();
 
 		if($this->csvHasHeader) {
 			$headers = array();
@@ -118,7 +116,10 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 			$fileData .= "\"" . implode("\"{$separator}\"", array_values($headers)) . "\"";
 			$fileData .= "\n";
 		}
-
+		
+		//Remove GridFieldPaginator as we're going to export the entire list.
+		$gridField->getConfig()->removeComponentsByType('GridFieldPaginator');
+		
 		$items = $gridField->getManipulatedList();
 
 		// @todo should GridFieldComponents change behaviour based on whether others are available in the config?
@@ -129,28 +130,37 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 		}
 
 		foreach($items->limit(null) as $item) {
-			$columnData = array();
+			if(!$item->hasMethod('canView') || $item->canView()) {
+				$columnData = array();
 
-			foreach($csvColumns as $columnSource => $columnHeader) {
-				if(!is_string($columnHeader) && is_callable($columnHeader)) {
-					if($item->hasMethod($columnSource)) {
-						$relObj = $item->{$columnSource}();
+				foreach($csvColumns as $columnSource => $columnHeader) {
+					if(!is_string($columnHeader) && is_callable($columnHeader)) {
+						if($item->hasMethod($columnSource)) {
+							$relObj = $item->{$columnSource}();
+						} else {
+							$relObj = $item->relObject($columnSource);
+						}
+
+						$value = $columnHeader($relObj);
 					} else {
-						$relObj = $item->relObject($columnSource);
+						$value = $gridField->getDataFieldValue($item, $columnSource);
+
+						if(!$value) {
+							$value = $gridField->getDataFieldValue($item, $columnHeader);
+						}
 					}
 
-					$value = $columnHeader($relObj);
-				} else {
-					$value = $gridField->getDataFieldValue($item, $columnSource);
+					$value = str_replace(array("\r", "\n"), "\n", $value);
+					$columnData[] = '"' . str_replace('"', '""', $value) . '"';
 				}
 
-				$value = str_replace(array("\r", "\n"), "\n", $value);
-				$columnData[] = '"' . str_replace('"', '\"', $value) . '"';
+				$fileData .= implode($separator, $columnData);
+				$fileData .= "\n";
 			}
-			$fileData .= implode($separator, $columnData);
-			$fileData .= "\n";
 
-			$item->destroy();
+			if($item->hasMethod('destroy')) {
+				$item->destroy();
+			}
 		}
 
 		return $fileData;
@@ -170,7 +180,7 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 		$this->exportColumns = $cols;
 		return $this;
 	}
-	
+
 	/**
 	 * @return string
 	 */
