+++ framework/forms/gridfield/GridFieldPrintButton.php
@@ -9,7 +9,7 @@
 class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionProvider, GridField_URLHandler {
 
 	/**
-	 * @var array Map of a property name on the printed objects, with values 
+	 * @var array Map of a property name on the printed objects, with values
 	 * being the column title in the CSV file.
 	 *
 	 * Note that titles are only used when {@link $csvHasHeader} is set to TRUE
@@ -20,7 +20,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 	 * @var boolean
 	 */
 	protected $printHasHeader = true;
-	
+
 	/**
 	 * Fragment to write the button to.
 	 *
@@ -46,10 +46,10 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 	 */
 	public function getHTMLFragments($gridField) {
 		$button = new GridField_FormAction(
-			$gridField, 
-			'print', 
+			$gridField,
+			'print',
 			_t('TableListField.Print', 'Print'),
-			'print', 
+			'print',
 			null
 		);
 
@@ -57,7 +57,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 		$button->addExtraClass('gridfield-button-print');
 
 		return array(
-			$this->targetFragment => '<p class="grid-print-button">' . $button->Field() . '</p>', 
+			$this->targetFragment => '<p class="grid-print-button">' . $button->Field() . '</p>',
 		);
 	}
 
@@ -115,7 +115,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 	 * Return the columns to print
 	 *
 	 * @param GridField
-	 * 
+	 *
 	 * @return array
 	 */
 	protected function getPrintColumnsForGridField(GridField $gridField) {
@@ -139,7 +139,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 	 */
 	public function getTitle(GridField $gridField) {
 		$form = $gridField->getForm();
-		$currentController = Controller::curr();
+		$currentController = $gridField->getForm()->getController();
 		$title = '';
 
 		if(method_exists($currentController, 'Title')) {
@@ -147,10 +147,8 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 		} else {
 			if ($currentController->Title) {
 				$title = $currentController->Title;
-			} else {
-				if($form->Name()){
-					$title = $form->Name();
-				}
+			} elseif ($form->getName()) {
+				$title = $form->getName();
 			}
 		}
 
@@ -172,7 +170,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
  	 */
 	public function generatePrintData(GridField $gridField) {
 		$printColumns = $this->getPrintColumnsForGridField($gridField);
-		
+
 		$header = null;
 
 		if($this->printHasHeader) {
@@ -184,26 +182,31 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 				)));
 			}
 		}
-		
+
 		$items = $gridField->getManipulatedList();
 		$itemRows = new ArrayList();
 
-		foreach($items as $item) {
+		foreach($items->limit(null) as $item) {
 			$itemRow = new ArrayList();
-			
+
 			foreach($printColumns as $field => $label) {
 				$value = $gridField->getDataFieldValue($item, $field);
+
+				if($item->escapeTypeForField($field) != 'xml') {
+					$value = Convert::raw2xml($value);
+				}
 
 				$itemRow->push(new ArrayData(array(
 					"CellString" => $value,
 				)));
 			}
-			
+
 			$itemRows->push(new ArrayData(array(
 				"ItemRow" => $itemRow
 			)));
-
-			$item->destroy();
+			if ($item->hasMethod('destroy')) {
+				$item->destroy();
+			}
 		}
 
 		$ret = new ArrayData(array(
@@ -213,7 +216,7 @@ class GridFieldPrintButton implements GridField_HTMLProvider, GridField_ActionPr
 			"Datetime" => SS_Datetime::now(),
 			"Member" => Member::currentUser(),
 		));
-		
+
 		return $ret;
 	}
